+++ src/QuickSort.java
@@ -1,5 +1,5 @@
 public class QuickSort implements SortingAlgorithm {
-    int basicOperations = 0;
+    long basicOperations = 0;
     int arrayLength = 0;
 
     @Override
@@ -44,7 +44,7 @@ public class QuickSort implements SortingAlgorithm {
     }
 
     @Override
-    public int getBasicOperations(){
+    public long getBasicOperations(){
         return basicOperations;
     }
 }
